requirements.txt
args==0.1.0
beautifulsoup4==4.6.3
bs4==0.0.1
clint==0.5.1
docopt==0.6.2
pkg-resources==0.0.0
python-dateutil==2.5.3
pyyaml>=4.2b1
requests>=2.20.0
schema==0.5.0
selenium==3.14.1
six==1.11.0
splinter==0.7.3
urljoin==1.0.0
urllib3>=1.24.2
wheel>=0.33.6

